Modify SiteFooter to use base components
